rdctl: Delete lima on factory reset
If the lima distribution exists, delete that on factory reset too.
