+++ src/go/rdctl/pkg/factoryreset/delete_data.go
@@ -22,9 +22,12 @@ import (
 	"fmt"
 	"io/fs"
 	"os"
+	"os/exec"
 	"path"
 
 	dockerconfig "github.com/docker/cli/cli/config"
+	"github.com/rancher-sandbox/rancher-desktop/src/go/rdctl/pkg/directories"
+	"github.com/rancher-sandbox/rancher-desktop/src/go/rdctl/pkg/paths"
 )
 
 type dockerConfigType map[string]any
@@ -85,4 +88,19 @@ func clearDockerContext() error {
 		return err
 	}
 	return os.Rename(scratchFile.Name(), configFilePath)
+}
+
+func deleteLimaVM() error {
+	appPaths, err := paths.GetPaths()
+	if err != nil {
+		return err
+	}
+	if err := directories.SetupLimaHome(appPaths.AppHome); err != nil {
+		return err
+	}
+	limactl, err := directories.GetLimactlPath()
+	if err != nil {
+		return err
+	}
+	return exec.Command(limactl, "delete", "-f", "0").Run()
 }
